refactor: rename directory utils to util
