+++ languages/golang/golang.go
@@ -21,9 +21,9 @@ import (
 	"gvm/internal/core"
 	"gvm/internal/http"
 	"gvm/internal/log"
-	"gvm/internal/utils/compress"
-	"gvm/internal/utils/path"
-	"gvm/internal/utils/slice"
+	"gvm/internal/util/compress"
+	"gvm/internal/util/path"
+	"gvm/internal/util/slice"
 	"gvm/languages"
 	"path/filepath"
 	"runtime"
